Change SetConfiguration parameter to short allowing for -1.
The libusb set_configuration documentation says that -1 should
be used to choose non-configured state for a device. The current
SetConfiguration uses a 'byte' type which is unsigned. Changing
to 'short' to allow for negative values.  Internally when the
USB message SET_CONFIGURATION is composed the value is typed as
'short' (at least in the .NET code).